+++ CHANGELOG.md
@@ -7,10 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
--   Add generated ngx-grpc client to package
+-   Add ngx-grpc client generation
 -   Add release checklist to README
--   Add opencontainers more opencontainers annotations to output server and envoy images
+-   Add more opencontainers annotations to output server and envoy images
 -   Remove clients package, copying the generated files would work better/be less complex
+-   Add `ghcr` example
 
 ## [0.2.2] - 2021-03-06