+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   Remove clients package, copying the generated files would work better/be less complex. Planning to replace this with an archive of client source files for each tag
 -   Add `ghcr` example
 -   Fix `tracing::instrument` not working in tonic async trait
+-   Start CSRF module development
 
 ## [0.2.2] - 2021-03-06